Vectors-of-known-length don't depend on Prelude.Linear anymore
Split Prelude.Linear into a minimalistic
part (Prelude.Linear.Internal.Simple) and the full
interface. Data.Vector.Linear now depends on
Prelude.Linear.Internal.Simple.
